Day 11 - Solutions 1+2
That was waaaay easier than expected. I feared a horrible exponential problem, which it WAS, but with
my approach it was really simple:
instead of really creating all stones, I just counted the nr of similar stones: because all stones
of the same nr act the same in a round, we can just calculate the new stone numbers for all same occurences.