Day 12 - Solutions 1+2
The first part was a simple flood-fill with some border tracking. Easy enough.
The 2nd part was trickier: First I didn't know how to count 'sides'. Then I realized,
that the numer of sides is the same as the number of corners of an area. So I tried
to detect all corners of one region, to get the nr of sides.
That worked, but the solution is somewhat ugly, see the code.